+++ src/app/actions/index.ts
@@ -18,6 +18,8 @@ export type { FindAdoptersInput, FindAdoptersOptions, FindAdoptersResponse, Dupl
 // Adopters
 export { getAdopter, saveAdopter, appendToExistingAdopter, getAdopterStats, logProfileView, getAverageRating, getHistory, getLinkedFormSubmissions, checkAdopterDeletable, deleteOwnAdopter, requestAdopterDeletion } from './adopters';
 export { addContactEntry } from './addContactEntry';
+export { updateContactEntry } from './updateContactEntry';
+export { removeContactEntry } from './removeContactEntry';
 
 // Images
 export { saveImage, getImages, setProfilePicture, getAdoptionImages, deleteImage } from './images';